backend-cms/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
bleach==6.1.0
blinker==1.8.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cssselect2==0.7.0
Django==5.0.7
django-classy-tags==4.1.0
django-cms==4.1.1
django-cors-headers==4.4.0
django-entangled==0.5.4
django-filer==3.1.3
django-formtools==2.5.1
django-fsm==2.8.2
django-parler==2.3
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangocms-alias==2.0.1
djangocms-attributes-field==3.0.0
djangocms-frontend==1.3.3
djangocms-text-ckeditor==5.1.5
djangocms-versioning==2.0.2
djangorestframework==3.15.2
easy-thumbnails==2.8.5
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
html5lib==1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pytz==2024.1
reportlab==4.2.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.3
